Refactor Flask app with blueprint factory
