+++ app/__init__.py
@@ -1,27 +1,36 @@
 from flask import Flask
-from .routes import bp_main, mysql
+from flask_mysqldb import MySQL
+
+mysql = MySQL()
+
+
+def crear_tablas():
+    cursor = mysql.connection.cursor()
+    cursor.execute('''CREATE TABLE IF NOT EXISTS usuarios (
+        id INT AUTO_INCREMENT PRIMARY KEY,
+        usuario VARCHAR(50) NOT NULL UNIQUE,
+        contrasena VARCHAR(100) NOT NULL
+    )''')
+    mysql.connection.commit()
+    cursor.close()
+
 
 def create_app():
     app = Flask(__name__)
     app.config.from_pyfile('config.py')
     app.secret_key = 'softgan_secret_key'
-    mysql.init_app(app)
-    app.register_blueprint(bp_main)
-    with app.app_context():
-        cursor = mysql.connection.cursor()
-        cursor.execute('''CREATE TABLE IF NOT EXISTS usuarios (
-            id INT AUTO_INCREMENT PRIMARY KEY,
-            usuario VARCHAR(50) NOT NULL UNIQUE,
-            contrasena VARCHAR(100) NOT NULL
-        )''')
-        mysql.connection.commit()
-        cursor.close()
-    return app
-    app = Flask(__name__)
-    app.config.from_pyfile('config.py')
-    app.register_blueprint(bp_main)
-    app.register_blueprint(bp_auth)
-    app.register_blueprint(bp_finca)
-    crear_tablas()
-    return app
 
+    mysql.init_app(app)
+
+    from .auth.routes import bp as auth_bp
+    from .ganaderia.routes import bp as ganaderia_bp
+    from .main.routes import bp as main_bp
+
+    app.register_blueprint(auth_bp)
+    app.register_blueprint(ganaderia_bp)
+    app.register_blueprint(main_bp)
+
+    with app.app_context():
+        crear_tablas()
+
+    return app
