+++ app/templates/dashboard.html
@@ -19,7 +19,7 @@
                     <div class="mb-2 text-sg-green fs-1"><i class="bi bi-droplet-half"></i></div>
                     <h5 class="card-title">Producción</h5>
                     <p class="card-text">Acceso rápido a registros de leche</p>
-                    <a href="{{ url_for('main.registro_leche') }}" class="btn btn-success w-100">Registrar Leche</a>
+                    <a href="{{ url_for('ganaderia.registro_leche') }}" class="btn btn-success w-100">Registrar Leche</a>
                 </div>
             </div>
         </div>
