+++ app/templates/login.html
@@ -31,7 +31,7 @@
       <button type="submit" class="login-btn">Ingresar</button>
     </form>
     <div class="mt-3 text-center">
-      <a href="{{ url_for('main.register') }}">¿No tienes cuenta? Regístrate</a>
+      <a href="{{ url_for('auth.register') }}">¿No tienes cuenta? Regístrate</a>
     </div>
   </div>
 </div>
