+++ app/templates/register.html
@@ -24,7 +24,7 @@
       <button type="submit" class="login-btn">Registrar</button>
     </form>
     <div class="mt-3 text-center">
-      <a href="{{ url_for('main.login') }}">Volver al login</a>
+      <a href="{{ url_for('auth.login') }}">Volver al login</a>
     </div>
   </div>
 </div>
